Fix some bugs with encode and compress -- still buggy.
The main bug right now is that the residuals can get much too
large (or perhaps m much too small) for the Rice encoding to be
effective. For example, with m = 6 and and residual of around
3,000, the Rice encoding will append 52 zeros to the end!
Unfortunately, many of the residuals are large like this
(hundreds of thousands for bass_1.wav). This makes the compressed
file extremely large (over 100 MB). Have not fixed THAT bug.